+++ homeassistant/components/daikin/sensor.py
