+++ homeassistant/components/homematicip_cloud/binary_sensor.py
