+++ homeassistant/components/homematicip_cloud/cover.py
