+++ homeassistant/components/homematicip_cloud/sensor.py
