+++ homeassistant/components/homematicip_cloud/switch.py
