+++ homeassistant/components/lifx/light.py
