+++ homeassistant/components/nest/binary_sensor.py
