+++ homeassistant/components/nest/camera.py
