+++ homeassistant/components/nest/sensor.py
