+++ homeassistant/components/point/binary_sensor.py
