+++ homeassistant/components/point/sensor.py
