+++ homeassistant/components/smhi/weather.py
