+++ homeassistant/components/tradfri/sensor.py
