+++ homeassistant/components/tradfri/switch.py
