+++ homeassistant/components/unifi/switch.py
