+++ homeassistant/components/upnp/sensor.py
