+++ tests/components/device_tracker/test_owntracks.py
@@ -1273,8 +1273,8 @@ async def test_single_waypoint_import(hass, context):
 
 async def test_not_implemented_message(hass, context):
     """Handle not implemented message type."""
-    patch_handler = patch('homeassistant.components.device_tracker.'
-                          'owntracks.async_handle_not_impl_msg',
+    patch_handler = patch('homeassistant.components.owntracks.'
+                          'device_tracker.async_handle_not_impl_msg',
                           return_value=mock_coro(False))
     patch_handler.start()
     assert not await send_message(hass, LWT_TOPIC, LWT_MESSAGE)
@@ -1283,8 +1283,8 @@ async def test_not_implemented_message(hass, context):
 
 async def test_unsupported_message(hass, context):
     """Handle not implemented message type."""
-    patch_handler = patch('homeassistant.components.device_tracker.'
-                          'owntracks.async_handle_unsupported_msg',
+    patch_handler = patch('homeassistant.components.owntracks.'
+                          'device_tracker.async_handle_unsupported_msg',
                           return_value=mock_coro(False))
     patch_handler.start()
     assert not await send_message(hass, BAD_TOPIC, BAD_MESSAGE)
@@ -1366,7 +1366,7 @@ def config_context(hass, setup_comp):
     patch_save.stop()
 
 
-@patch('homeassistant.components.device_tracker.owntracks.get_cipher',
+@patch('homeassistant.components.owntracks.device_tracker.get_cipher',
        mock_cipher)
 async def test_encrypted_payload(hass, config_context):
     """Test encrypted payload."""
@@ -1377,7 +1377,7 @@ async def test_encrypted_payload(hass, config_context):
     assert_location_latitude(hass, LOCATION_MESSAGE['lat'])
 
 
-@patch('homeassistant.components.device_tracker.owntracks.get_cipher',
+@patch('homeassistant.components.owntracks.device_tracker.get_cipher',
        mock_cipher)
 async def test_encrypted_payload_topic_key(hass, config_context):
     """Test encrypted payload with a topic key."""
@@ -1390,7 +1390,7 @@ async def test_encrypted_payload_topic_key(hass, config_context):
     assert_location_latitude(hass, LOCATION_MESSAGE['lat'])
 
 
-@patch('homeassistant.components.device_tracker.owntracks.get_cipher',
+@patch('homeassistant.components.owntracks.device_tracker.get_cipher',
        mock_cipher)
 async def test_encrypted_payload_no_key(hass, config_context):
     """Test encrypted payload with no key, ."""
@@ -1403,7 +1403,7 @@ async def test_encrypted_payload_no_key(hass, config_context):
     assert hass.states.get(DEVICE_TRACKER_STATE) is None
 
 
-@patch('homeassistant.components.device_tracker.owntracks.get_cipher',
+@patch('homeassistant.components.owntracks.device_tracker.get_cipher',
        mock_cipher)
 async def test_encrypted_payload_wrong_key(hass, config_context):
     """Test encrypted payload with wrong key."""
@@ -1414,7 +1414,7 @@ async def test_encrypted_payload_wrong_key(hass, config_context):
     assert hass.states.get(DEVICE_TRACKER_STATE) is None
 
 
-@patch('homeassistant.components.device_tracker.owntracks.get_cipher',
+@patch('homeassistant.components.owntracks.device_tracker.get_cipher',
        mock_cipher)
 async def test_encrypted_payload_wrong_topic_key(hass, config_context):
     """Test encrypted payload with wrong  topic key."""
@@ -1427,7 +1427,7 @@ async def test_encrypted_payload_wrong_topic_key(hass, config_context):
     assert hass.states.get(DEVICE_TRACKER_STATE) is None
 
 
-@patch('homeassistant.components.device_tracker.owntracks.get_cipher',
+@patch('homeassistant.components.owntracks.device_tracker.get_cipher',
        mock_cipher)
 async def test_encrypted_payload_no_topic_key(hass, config_context):
     """Test encrypted payload with no topic key."""
